Improve Binary Serialization in the Code Generator
I'm a big fan of the Code Generator tool, but I noticed the all-in-one pages serialize the game as a JSON array, which can get kind of chunky.

This PR swaps the JSON for a base64 string, bringing a `53.35mb` .html down to `21.85mb`.   This should help with storage and load-times.

Zipping the .html afterwards brings the whole thing down to within 10% of the original game size.

bytes <-> base64 conversions adopted from examples on this page: https://developer.mozilla.org/en-US/docs/Web/API/Window/btoa