Add Compression to Code Generator
This PR adds compression to the generated .html files using the browser's built-in gzip compression/decompression routines.

This brings what would have been a `53.35mb` .html on master (or a `21.85mb` .html with #58 ) down to `16.32mb`.   This further compresses down to `12mb` if zipped.

Due to the async call needed for this, I made this a separate PR since the structure of the generated .html has changed to only load the emulator after the game data has been decompressed.   This is incompatible with the non-single-file-packing code, so it has been removed.